Moved calibration to its own module
