+++ pyobs/images/processors/misc/__init__.py
@@ -4,7 +4,7 @@ Misc
 """
 
 from .addmask import AddMask
-from .calibration import Calibration
+from .calibration.calibration import Calibration
 from .createfilename import CreateFilename
 from .removebackground import RemoveBackground
 from .smooth import Smooth
