+++ pyobs/images/processors/misc/calibration/calibration.py
@@ -1,13 +1,10 @@
-import asyncio
-from functools import partial
-from typing import Union, Optional, List, Tuple, Any, Dict, cast
+from typing import Union, Optional, Tuple, Any, Dict, cast
 import logging
-import astropy.units as u
 
 from pyobs.images.processor import ImageProcessor
 from pyobs.images import Image
-from pyobs.images.processors.misc._calibration_cache import _CalibrationCache
-from pyobs.images.processors.misc._ccddata_calibrator import _CCDDataCalibrator
+from pyobs.images.processors.misc.calibration._calibration_cache import _CalibrationCache
+from pyobs.images.processors.misc.calibration._ccddata_calibrator import _CCDDataCalibrator
 from pyobs.object import get_object
 from pyobs.utils.archive import Archive
 from pyobs.utils.enums import ImageType
